Bug: Error message shown after successfully deleting content.
This issue occurred because, after successfully deleting the content, the module switched to the deletionSuccess action, which doesn’t exist (nor does the deletionError action). We’ve replaced these with existing actions to fix the error.action, wich doesn't exists, neither the deletionError action, instead we replaced those for actions that actually exists to fix the error.